Split syntax rewriters into individual files
