+++ RefactorMCP.ConsoleApp/SyntaxRewriters/InstanceMemberRewriter.cs
@@ -1,0 +1,46 @@
+using Microsoft.CodeAnalysis;
+using Microsoft.CodeAnalysis.CSharp;
+using Microsoft.CodeAnalysis.CSharp.Syntax;
+using Microsoft.CodeAnalysis.Text;
+using System.Collections.Generic;
+using System.Linq;
+
+internal class InstanceMemberRewriter : CSharpSyntaxRewriter
+{
+    private readonly string _parameterName;
+    private readonly HashSet<string> _knownInstanceMembers;
+
+    public InstanceMemberRewriter(string parameterName, HashSet<string> knownInstanceMembers)
+    {
+        _parameterName = parameterName;
+        _knownInstanceMembers = knownInstanceMembers;
+    }
+
+    public override SyntaxNode? VisitIdentifierName(IdentifierNameSyntax node)
+    {
+        var parent = node.Parent;
+        if (parent is ParameterSyntax || parent is TypeSyntax)
+            return base.VisitIdentifierName(node);
+
+        if (_knownInstanceMembers.Contains(node.Identifier.ValueText))
+        {
+            if (parent is MemberAccessExpressionSyntax memberAccess && memberAccess.Expression == node)
+            {
+                return SyntaxFactory.MemberAccessExpression(
+                    SyntaxKind.SimpleMemberAccessExpression,
+                    SyntaxFactory.IdentifierName(_parameterName),
+                    node);
+            }
+            else if (parent is not MemberAccessExpressionSyntax)
+            {
+                return SyntaxFactory.MemberAccessExpression(
+                    SyntaxKind.SimpleMemberAccessExpression,
+                    SyntaxFactory.IdentifierName(_parameterName),
+                    node);
+            }
+        }
+
+        return base.VisitIdentifierName(node);
+    }
+}
+
