+++ RefactorMCP.ConsoleApp/SyntaxRewriters/InstanceMemberUsageChecker.cs
@@ -1,0 +1,39 @@
+using Microsoft.CodeAnalysis;
+using Microsoft.CodeAnalysis.CSharp;
+using Microsoft.CodeAnalysis.CSharp.Syntax;
+using Microsoft.CodeAnalysis.Text;
+using System.Collections.Generic;
+using System.Linq;
+
+internal class InstanceMemberUsageChecker : CSharpSyntaxRewriter
+{
+    private readonly HashSet<string> _knownInstanceMembers;
+    public bool HasInstanceMemberUsage { get; private set; }
+
+    public InstanceMemberUsageChecker(HashSet<string> knownInstanceMembers)
+    {
+        _knownInstanceMembers = knownInstanceMembers;
+    }
+
+    public override SyntaxNode? VisitIdentifierName(IdentifierNameSyntax node)
+    {
+        var parent = node.Parent;
+        if (parent is ParameterSyntax || parent is TypeSyntax)
+            return base.VisitIdentifierName(node);
+
+        if (_knownInstanceMembers.Contains(node.Identifier.ValueText))
+        {
+            if (parent is MemberAccessExpressionSyntax memberAccess && memberAccess.Expression == node)
+            {
+                HasInstanceMemberUsage = true;
+            }
+            else if (parent is not MemberAccessExpressionSyntax && parent is not InvocationExpressionSyntax)
+            {
+                HasInstanceMemberUsage = true;
+            }
+        }
+
+        return base.VisitIdentifierName(node);
+    }
+}
+
