+++ RefactorMCP.ConsoleApp/SyntaxRewriters/MethodCallChecker.cs
@@ -1,0 +1,31 @@
+using Microsoft.CodeAnalysis;
+using Microsoft.CodeAnalysis.CSharp;
+using Microsoft.CodeAnalysis.CSharp.Syntax;
+using Microsoft.CodeAnalysis.Text;
+using System.Collections.Generic;
+using System.Linq;
+
+internal class MethodCallChecker : CSharpSyntaxRewriter
+{
+    private readonly HashSet<string> _classMethodNames;
+    public bool HasMethodCalls { get; private set; }
+
+    public MethodCallChecker(HashSet<string> classMethodNames)
+    {
+        _classMethodNames = classMethodNames;
+    }
+
+    public override SyntaxNode? VisitInvocationExpression(InvocationExpressionSyntax node)
+    {
+        if (node.Expression is IdentifierNameSyntax identifier)
+        {
+            if (_classMethodNames.Contains(identifier.Identifier.ValueText))
+            {
+                HasMethodCalls = true;
+            }
+        }
+
+        return base.VisitInvocationExpression(node);
+    }
+}
+
